requirements.txt
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.2
django-filter==1.0.4
django-filters==0.2.1
future==0.16.0
gunicorn==19.7.1
idna==2.5
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
python-decouple==3.0
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
viberbot==1.0.11
whitenoise==3.3.0
